+++ RELEASENOTES.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ## Unreleased
+- Add support for Subject Name + Issuer (SNI) certificate authentication ([#46 by philon-msft](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/46))
 
 ## 2.0.0
 - Fix [#2](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/issues/2), [#25](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/issues/25): Add TokenCredential/DefaultAzureCredential support ([#10 by lsannicolas](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/10))